chore: create constant file
