+++ constants.ts
@@ -1,0 +1,11 @@
+// constant to record colors that map to tailwind tokens
+export const COLORS = [
+  { name: 'red', class: 'bg-red-200' },
+  { name: 'blue', class: 'bg-blue-200' },
+  { name: 'green', class: 'bg-green-200' },
+  { name: 'yellow', class: 'bg-yellow-200' },
+  { name: 'purple', class: 'bg-purple-200' },
+  { name: 'pink', class: 'bg-pink-200' },
+  { name: 'orange', class: 'bg-orange-200' },
+  { name: 'gray', class: 'bg-gray-200' },
+];
